+++ src/main/java/com/example/springboot/services/impl/UserServiceImpl.java
@@ -1,8 +1,11 @@
 package com.example.springboot.services.impl;
 
 import com.example.springboot.entities.User;
+import com.example.springboot.exceptions.ItemNotFoundException;
 import com.example.springboot.repositories.UserRepository;
 import com.example.springboot.services.UserService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,10 +13,11 @@ import org.springframework.transaction.annotation.Transactional;
 import java.time.LocalDate;
 import java.util.List;
 
+@Slf4j
+@RequiredArgsConstructor
 @Service
 @Transactional
 public class UserServiceImpl implements UserService {
-    @Autowired
     UserRepository userRepository;
 
     @Override
@@ -29,13 +33,15 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User getByUsername(String username) {
-        return userRepository.findByUsername(username);
+    public User getByUsername(String username) throws ItemNotFoundException {
+        return userRepository.findByUsername(username)
+                .orElseThrow(() -> new ItemNotFoundException("utente non trovato"));
     }
 
     @Override
-    public User getByUsernameAndPassword(String username, String password) {
-        return userRepository.findByUsernameAndPassword(username, password);
+    public User getByUsernameAndPassword(String username, String password) throws ItemNotFoundException {
+        return userRepository.findByUsernameAndPassword(username, password)
+                .orElseThrow(() -> new ItemNotFoundException("username o password errati"));
     }
 
     @Override
@@ -44,7 +50,14 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void delUser(User user) {
+    public void delUser(long id) throws ItemNotFoundException {
+        log.info("eliminazione user");
+
+        User user = getById(id);
+
+        if (user == null) {
+            throw new ItemNotFoundException("utente non trovato");
+        }
         userRepository.delete(user);
     }
